Organized all the structures of the folders and files, renaming them to be more intuitive
