--- tmp_tatchi/Renergy_Growth.ipynb:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-HostUrl=https://files.slack.com/files-pri/T02NE0241-F05BYRFKPE0/download/renergy_growth.ipynb?origin_team=T02NE0241
