--- tmp_tatchi/full_cleaned_final.csv:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-HostUrl=https://files.slack.com/files-pri/T02NE0241-F05C921GYQ1/download/full_cleaned_final.csv?origin_team=T02NE0241
