--- tmp_tatchi/preprocessing_final.ipynb:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-HostUrl=https://files.slack.com/files-pri/T02NE0241-F05BPNVJ7GW/download/preprocessing_final.ipynb?origin_team=T02NE0241
