--- tmp_tatchi/preprocessing_final_1.ipynb:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-HostUrl=https://files.slack.com/files-pri/T02NE0241-F05BW936F3M/download/preprocessing_final_1.ipynb?origin_team=T02NE0241
